# Add basic precompilation workload