@ablaom 2024-06-06
Closing in favour of adding workloads added to component packages. See also https://github.com/JuliaAI/MLJBase.jl/issues/924